refactoring AccountService.java

+++ src/main/java/org/das/service/AccountService.java
@@ -2,12 +2,13 @@ package org.das.service;
 
 import org.das.model.Account;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface AccountService {
     Account accountCreate(UUID userId);
     void accountClose(UUID accountId);
-    void accountDeposit(UUID accountId, double amount);
-    void accountTransfer(UUID senderId, UUID recipientId, double amount);
-    void accountWithdraw(UUID accountId, double amount);
+    void accountDeposit(UUID accountId, BigDecimal amount);
+    void accountTransfer(UUID senderId, UUID recipientId, BigDecimal amount);
+    void accountWithdraw(UUID accountId, BigDecimal amount);
 }
